requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.3.1
facenet-pytorch==2.6.0
fastapi==0.135.1
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
opencv-python-headless==4.11.0.86
packaging==26.0
pandas==2.3.3
pillow==10.2.0
platformdirs==4.9.4
protobuf==6.33.5
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.7.2
scipy==1.15.3
sentry-sdk==2.54.0
six==1.17.0
smmap==5.0.3
starlette==0.52.1
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
wandb==0.25.1
grad-cam==1.5.5
python-multipart==0.0.9
